requirements.txt
Django==2.2.2
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
django-cors-headers==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
